timeline: Support Block API Version 2
