+++ blocks/src/block/timeline-item/block.json
@@ -1,0 +1,20 @@
+{
+  "apiVersion": 2,
+  "name": "cocoon-blocks/timeline-item",
+  "category": "cocoon-block",
+  "parent": ["cocoon-blocks/timeline"],
+
+  "attributes": {
+    "label": {
+      "type": "string",
+      "default": "ラベル"
+    },
+    "title": {
+      "type": "string",
+      "default": "タイトル"
+    }
+  },
+  "supports": {
+    "inserter": false
+  }
+}
