+++ blocks/src/block/timeline/block.json
@@ -1,0 +1,38 @@
+{
+  "apiVersion": 2,
+  "name": "cocoon-blocks/timeline",
+  "category": "cocoon-block",
+  "keywords": [ "timeline", "box" ],
+
+  "attributes": {
+    "title": {
+      "type": "string",
+      "default": "タイムラインのタイトル"
+    },
+    "items": {
+      "type": "integer",
+      "default": 1
+    },
+    "backgroundColor": {
+      "type": "string"
+    },
+    "textColor": {
+      "type": "string"
+    },
+    "borderColor": {
+      "type": "string"
+    },
+    "customBorderColor": {
+      "type": "string"
+    },
+    "pointColor": {
+      "type": "string"
+    },
+    "customPointColor": {
+      "type": "string"
+    },
+    "fontSize": {
+      "type": "string"
+    }
+  }
+}
