+++ blocks/src/init.php
@@ -286,4 +286,6 @@ require_once abspath(__FILE__).'block/info-box/index.php';
 require_once abspath(__FILE__).'block/search-box/index.php';
 require_once abspath(__FILE__).'block/sticky-box/index.php';
 require_once abspath(__FILE__).'block/tab-box/index.php';
+require_once abspath(__FILE__).'block/timeline/index.php';
+require_once abspath(__FILE__).'block/timeline-item/index.php';
 require_once abspath(__FILE__).'block/toggle-box/index.php';
